fixed problems with deleting and updating tasks

+++ project_django/task/todolist_app/views.py
@@ -31,7 +31,12 @@ def todolist(request):
 @login_required
 def delete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.delete()
+    if task.manage == request.user:
+        task.delete()
+    else:
+        messages.error(request,("Access Restricted, You Are Not Allowed!"))
+   
+    
     return redirect('todolist')
 
 @login_required
@@ -51,8 +56,11 @@ def update_task(request, task_id):
 @login_required
 def complete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.done = True
-    task.save()
+    if task.manage == request.user:
+        task.done = True
+        task.save()
+    else:
+        messages.error(request,("Access Restricted, You Are Not Allowed!"))
     return redirect('todolist')
 
 @login_required
